+++ Rebus/Config/RebusConfigurer.cs
@@ -86,7 +86,7 @@ namespace Rebus.Config
         public RebusConfigurer DataBus(Action<StandardConfigurer<IDataBusStorage>> configurer)
         {
             if (configurer == null) throw new ArgumentNullException(nameof(configurer));
-            
+
             configurer(new StandardConfigurer<IDataBusStorage>(_injectionist, _options));
 
             if (_injectionist.Has<IDataBusStorage>())
@@ -203,7 +203,7 @@ namespace Rebus.Config
             _injectionist.Register(c => c.Get<CancellationTokenSource>().Token);
 
             PossiblyRegisterDefault<IRebusLoggerFactory>(c => new ConsoleLoggerFactory(true));
-            
+
             PossiblyRegisterDefault<IRebusTime>(c => new DefaultRebusTime());
 
             //PossiblyRegisterDefault<IAsyncTaskFactory>(c => new TimerAsyncTaskFactory(c.Get<IRebusLoggerFactory>()));
@@ -225,7 +225,7 @@ namespace Rebus.Config
 
             PossiblyRegisterDefault<ITimeoutManager>(c => new DisabledTimeoutManager());
 
-            PossiblyRegisterDefault<ISerializer>(c => new JsonSerializer());
+            PossiblyRegisterDefault<ISerializer>(c => new JsonSerializer(c.Get<IMessageTypeNameConvention>()));
 
             PossiblyRegisterDefault<IPipelineInvoker>(c =>
             {
@@ -332,6 +332,7 @@ namespace Rebus.Config
                 var rebusLoggerFactory = c.Get<IRebusLoggerFactory>();
                 var options = c.Get<Options>();
                 var rebusTime = c.Get<IRebusTime>();
+                var messageTypeNameConvention = c.Get<IMessageTypeNameConvention>();
 
                 return new DefaultPipeline()
                     .OnReceive(c.Get<IRetryStrategyStep>())
@@ -344,7 +345,7 @@ namespace Rebus.Config
                     .OnReceive(new LoadSagaDataStep(c.Get<ISagaStorage>(), rebusLoggerFactory))
                     .OnReceive(new DispatchIncomingMessageStep(rebusLoggerFactory))
 
-                    .OnSend(new AssignDefaultHeadersStep(transport, rebusTime, options.DefaultReturnAddressOrNull))
+                    .OnSend(new AssignDefaultHeadersStep(transport, rebusTime, messageTypeNameConvention, options.DefaultReturnAddressOrNull))
                     .OnSend(new FlowCorrelationIdStep())
                     .OnSend(new AutoHeadersOutgoingStep())
                     .OnSend(new SerializeOutgoingMessageStep(serializer))
@@ -359,6 +360,8 @@ namespace Rebus.Config
             PossiblyRegisterDefault<IDataBus>(c => new DisabledDataBus());
 
             PossiblyRegisterDefault<ITopicNameConvention>(c => new DefaultTopicNameConvention());
+
+            PossiblyRegisterDefault<IMessageTypeNameConvention>(c => new SimpleAssemblyQualifiedMessageTypeNameConvention());
 
             // configuration hack - keep these two bad boys around to have them available at the last moment before returning the built bus instance...
             Action startAction = null;